Add script to populate series information
The series name and ordering of badges is determined by badge name. It
is based on heuristic with badge name: if the name ends with
parenthesised expression with a number, it is parsed and put into the
database.

If the series with required name does not exist yet, a notice is printed
and nothing more happens.